pyproject.toml
[project]
name = "python-project-template"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.25",
    "langchain-openai>=0.3.19",
    "langchain-sandbox>=0.0.4",
    "langgraph>=0.3.34",
    "langgraph-codeact>=0.1.3",
    "numpy>=2.2.6",
    "python-dotenv>=1.1.0",
]

[tool.mypy]
strict = true

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D", "ANN", "TRY", "EM", "FA", "SIM108", "S101", "RET505", "TCH002"]
